Semana 12
Se completaron las funciones de solicitud y se inicio la funcion de
busqueda.

+++ views/libro/cargarLibros.php
@@ -3,6 +3,20 @@
 //die;
 ?>
 <center>
+    <!--BUSCADOR-->
+    <div class="col-xs-4">
+        Búsqueda por identificación:
+    </div>
+    <div class="col-xs-2">
+        <input type="text" class="input-sm validate[required]" name="txt_nombreLibro" id="txt_nombreLibro" />
+    </div>
+    <div class="col-xs-2">
+        <input type="button" class="btn-sm btn-success" id="buscarLibroRatificar" value="Buscar" />
+    </div>
+    <div class="col-xs-offset-5"></div>
+    <div class="col-xs-12"><br></div>
+    <div class="col-lg-10 col-xs-12"></div>
+    <!--TABLA-->
     <table class="table table-condensed">
         <tr>
             <th colspan="8" class="nombreTabla text-center">Lista de Libros</th>
@@ -13,7 +27,12 @@
             <th>Autor</th>
             <th>Categoría</th>
             <th>Editorial</th>
-            <th colspan="2" class="text-center">Acción</th>
+            <?php if (Session::get('tipoUsuario') < 1) { ?>
+                <th colspan="2" class="text-center">Acción</th>
+            <?php } ?>
+            <?php if (Session::get('tipoUsuario') >= 1) { ?>
+                <th colspan="2" class="text-center"></th>
+            <?php } ?>
         </tr>
         <?php
         $con = 1;
@@ -36,8 +55,10 @@
             echo $value['editorial'];
             echo '</td>';
             echo '<td class = text-center>';
-            echo '<a class="btn-sm btn-warning" href="editarLibro/' . $value['codigo'] . '">Editar</a> &nbsp; &nbsp; &nbsp;';
-            echo '<a class="btn-sm btn-danger" href="eliminarLibro/' . $value['codigo'] . '" onclick = "return confirm(' . $mensaje . ');">Eliminar</a>';
+            if (Session::get('tipoUsuario') < 1) {
+                echo '<a class="btn-sm btn-warning" href="editarLibro/' . $value['codigo'] . '">Editar</a> &nbsp; &nbsp; &nbsp;';
+                echo '<a class="btn-sm btn-danger" href="eliminarLibro/' . $value['codigo'] . '" onclick = "return confirm(' . $mensaje . ');">Eliminar</a>';
+            }
             echo '</td>';
             echo '</tr>';
             $con++;
